LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-size-shadow-invalidation-expected.txt

Green
Green

FAIL Host container child invalidated with container in shadow tree assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Non-host container child invalidated with container in shadow tree assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
